Load user principal from cookie
Always return an authentication success, if visitor is not logged
in, we return the "Anonymous User" principal.  Added a validation
that no new member name can match the anonymous user name.

Converted a couple value objects to use the new "of" idiom.  This
is prep for pushing command validation into the commands, which
should enable more natural railroad-style validation pipelines.