Brute force implementation of one security policy
Introduce the Guard type, which is a tuple of two predicates.  The first
returns true if the method matches a resource, and the second returns
true if the subject matches a condition.

For example, the guard (isCreateArticle(Method), isLoggedIn(Subject))
implements a single entitlement.

Right now, the mapping from entitlement to guard is hard code to this
specific guard.

Also, configure IntelliJ to suppress warnings about comments in JSON, as
gson supports this out of the box.