Add test for a diary note post
Add from date to note post so we put it in the right month.  Return the
content id of the newly created note in the POST response.  Rename the
resource from "add-note" to the more typical "notes".

Move normalizeDates from PhotoHandler into FormData.  Change the names of
the date form field to be general enough to cover both photos and notes.